Move packages, add css classes
