+++ jvm/markerapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/markerapp/app/view/layers/Timecode.kt
@@ -1,10 +1,12 @@
-package org.wycliffeassociates.otter.jvm.markerapp.app.view
+package org.wycliffeassociates.otter.jvm.markerapp.app.view.layers
 
 import javafx.scene.SnapshotParameters
 import javafx.scene.canvas.Canvas
 import javafx.scene.image.Image
 import javafx.scene.image.WritableImage
 import javafx.scene.paint.Color
+import org.wycliffeassociates.otter.jvm.markerapp.app.view.pixelsInSecond
+import org.wycliffeassociates.otter.jvm.markerapp.app.view.positionToMs
 import tornadofx.*
 import java.lang.Math.floor
 import java.util.concurrent.TimeUnit
