+++ jvm/markerapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/markerapp/app/view/layers/TimecodeHolder.kt
@@ -3,7 +3,6 @@ package org.wycliffeassociates.otter.jvm.markerapp.app.view.layers
 import com.sun.glass.ui.Screen
 import javafx.geometry.Rectangle2D
 import javafx.scene.image.ImageView
-import org.wycliffeassociates.otter.jvm.markerapp.app.view.Timecode
 import org.wycliffeassociates.otter.jvm.markerapp.app.viewmodel.VerseMarkerViewModel
 import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNow
